feat: table_routes util
